pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zhengjie"
version = "0.1.0"
description = "A production-oriented LangChain agent service template."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=1.5.5",
  "fastapi>=0.115,<1.0",
  "langchain>=1.0,<2.0",
  "langchain-anthropic>=1.4.0",
  "langchain-ollama>=1.0.1",
  "langchain-openai>=1.0,<2.0",
  "langgraph>=1.1.3",
  "python-dotenv>=1.0,<2.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["src*"]

